backend/requirements.txt
livekit-agents[silero,turn-detector,google,deepgram,cartesia,elevenlabs,assemblyai]
livekit-plugins-noise-cancellation
google-generativeai
python-dotenv

fastapi
uvicorn
livekit